+++ core/domain/interface/merchant/merchant.go
@@ -46,7 +46,7 @@ type (
 		SelfSales() bool
 
 		// 返回对应的会员编号
-		Member() int
+		Member() int64
 
 		// 保存
 		Save() (int64, error)
@@ -97,7 +97,7 @@ type (
 		Save() error
 
 		// 根据编号获取余额变动信息
-		GetBalanceLog(id int) *BalanceLog
+		GetBalanceLog(id int64) *BalanceLog
 
 		// 根据号码获取余额变动信息
 		GetBalanceLogByOuterNo(outerNo string) *BalanceLog
@@ -126,16 +126,16 @@ type (
 		CreateSignUpToken(memberId int64) string
 
 		// 根据商户申请密钥获取会员编号
-		GetMemberFromSignUpToken(token string) int
+		GetMemberFromSignUpToken(token string) int64
 
 		// 提交商户注册信息
-		CommitSignUpInfo(*MchSignUp) (int, error)
+		CommitSignUpInfo(*MchSignUp) (int64, error)
 
 		// 审核商户注册信息
-		ReviewMchSignUp(id int, pass bool, remark string) error
+		ReviewMchSignUp(id int64, pass bool, remark string) error
 
 		// 获取商户申请信息
-		GetSignUpInfo(id int) *MchSignUp
+		GetSignUpInfo(id int64) *MchSignUp
 
 		// 获取会员申请的商户信息
 		GetSignUpInfoByMemberId(memberId int64) *MchSignUp
@@ -149,11 +149,11 @@ type (
 
 	// 商户申请信息
 	MchSignUp struct {
-		Id int `db:"id" pk:"yes" auth:"yes"`
+		Id int64 `db:"id" pk:"yes" auth:"yes"`
 		// 申请单号
 		SignNo string `db:"sign_no"`
 		// 会员编号
-		MemberId int `db:"member_id"`
+		MemberId int64 `db:"member_id"`
 		// 用户名
 		Usr string `db:"usr"`
 		// 密码
@@ -161,11 +161,11 @@ type (
 		// 商户名称号
 		MchName string `db:"mch_name"`
 		// 省
-		Province int `db:"province"`
+		Province int64 `db:"province"`
 		// 市
-		City int `db:"city"`
+		City int64 `db:"city"`
 		// 区
-		District int `db:"district"`
+		District int64 `db:"district"`
 		// 详细地址
 		Address string `db:"address"`
 		// 店铺店铺
@@ -198,9 +198,9 @@ type (
 
 	// 商户
 	Merchant struct {
-		Id int `db:"id" pk:"yes" auto:"yes"`
+		Id int64 `db:"id" pk:"yes" auto:"yes"`
 		// 关联的会员编号,作为结算账户
-		MemberId int `db:"member_id"`
+		MemberId int64 `db:"member_id"`
 		// 用户
 		Usr string `db:"usr"`
 		// 密码
@@ -214,11 +214,11 @@ type (
 		// 标志
 		Logo string `db:"logo"`
 		// 省
-		Province int `db:"province"`
+		Province int64 `db:"province"`
 		// 市
-		City int `db:"city"`
+		City int64 `db:"city"`
 		// 区
-		District int `db:"district"`
+		District int64 `db:"district"`
 		// 是否启用
 		Enabled int `db:"enabled"`
 		// 过期时间
@@ -236,7 +236,7 @@ type (
 	// 商户账户表
 	Account struct {
 		// 商户编号
-		MchId int `db:"mch_id" pk:"yes"`
+		MchId int64 `db:"mch_id" pk:"yes"`
 		// 余额
 		Balance float32 `db:"balance"`
 		// 冻结金额
@@ -260,9 +260,9 @@ type (
 	// 商户余额日志
 	BalanceLog struct {
 		// 编号
-		Id int `db:"id" pk:"yes" auto:"yes"`
+		Id int64 `db:"id" pk:"yes" auto:"yes"`
 		// 商户编号
-		MchId int `db:"mch_id"`
+		MchId int64 `db:"mch_id"`
 		// 日志类型
 		Kind int `db:"kind"`
 		// 标题
@@ -284,9 +284,9 @@ type (
 	// 商户每日报表
 	MchDayChart struct {
 		// 编号
-		Id int `db:"id" pk:"yes" auto:"yes"`
+		Id int64 `db:"id" pk:"yes" auto:"yes"`
 		// 商户编号
-		MchId int `db:"mch_id"`
+		MchId int64 `db:"mch_id"`
 		// 新增订单数量
 		OrderNumber int `db:"order_number"`
 		// 订单额
