+++ core/repository/value_rep.go
@@ -344,7 +344,7 @@ func (vp *valueRep) GetAreaNames(id []int) []string {
 }
 
 // 获取省市区字符串
-func (vp *valueRep) GetAreaString(province, city, district int) string {
+func (vp *valueRep) GetAreaString(province, city, district int64) string {
 	names := vp.GetAreaNames([]int{province, city, district})
 	if names[1] == "市辖区" || names[1] == "市辖县" || names[1] == "县" {
 		return strings.Join([]string{names[0], names[2]}, " ")
